Fixed navigation for student side.
