+++ react-app/src/screens/Host/HostManage.js
@@ -13,7 +13,6 @@ import {
 
 import { AntDesign } from "@expo/vector-icons";
 import { Header } from "react-native-elements";
-import eventsData from "../../assets/events-data/eventsData";
 import React, { useEffect } from "react";
 import { TextInput } from "react-native";
 import { FlatList } from "react-native-gesture-handler";
@@ -188,6 +187,7 @@ export default function HostManage({ navigation, route }) {
       "createEventHost",
       { hid: loginState.id, newEvent: newEvent },
       (err, res) => {
+        console.log(err, res);
         if (err) {
           Alert.alert("Error", "Could not create new event.", [
             {
@@ -251,7 +251,7 @@ export default function HostManage({ navigation, route }) {
   );
 
   const renderItem = ({ item }) => {
-    return <Item item={item} onPress={() => {}} />;
+    return <Item item={item} onPress={() => { }} />;
   };
 
   return (
