+++ react-app/src/screens/Host/editEvents.js
